fix: wrong path of test folder in Java Interpreter
